requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
requests>=2.31.0
twilio>=9.0.0
python-dotenv>=1.0.0